Fixed loading player at transport. Fixed type conversion errors and unused method in header.
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@2629 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/Transports.h
@@ -33,7 +33,7 @@ public:
 private:
     struct WayPoint {
         WayPoint() : mapid(0), x(0), y(0), z(0), teleport(0) {}
-        WayPoint(float _mapid, float _x, float _y, float _z, bool _teleport) : 
+        WayPoint(uint32 _mapid, float _x, float _y, float _z, bool _teleport) : 
                 mapid(_mapid), x(_x), y(_y), z(_z), teleport(_teleport) {}
         uint32 mapid;
         float x;
